Edge case - audio settings
Reads audio settings in a csv file on the USB stick (every time it's plugged).
If there's no audio_settings.csv file, it creates one with the default settings (written in the program).
Among the audio settings, cutoff and threshold can be configured, but for now it's not possible to set a device as input.
The edge cases related to storage device and microphone are correctly handled.
(NB: too many functions in main.py)